improved installer

+++ installer.py
@@ -1,28 +1,63 @@
 import urllib2
 #    http://stackoverflow.com/questions/24346872/python-equivalent-of-a-given-wget-command
-import os,shutil,sys
+import os,shutil,sys,re
 import zipfile
 import pickle
 import datetime
 import argparse
 import importlib
 import traceback
+from argparse import Namespace
 
 #constants:
 MINIMAL_WX_VERSION       = (2,8,10)
 MINIMAL_PARAMIKO_VERSION = (1,15,2)
 ZIP_FILE = 'Launcher.zip'
 UNZIPPED_FOLDER_PREFIX = "hpcuantwerpen-Launcher-"
-GIT_COMMIT_ID_PREFIX   = "git_commit_id_"
+GIT_COMMIT_ID_PREFIX   = "Global.git_commit_id_"
 LAUNCHER_PREVIOUS      = "Launcher.bak"
+GLOBAL_PICKLED         = "installer_status.pickled"
 
-#global variables:
-user_home     = None
-launcher_home = None
-istep  = 0
-unzipped_folder = None
-git_commit_id = None
-dependencies_ok = False
+class Global(object): 
+    """ using class as namespace
+        key-value pairs are defined in loadGlobal()
+    """
+
+def loadGlobal(reset=False):
+    """ persistence for Global namespace"""
+    if reset:
+        #set default key value pairs
+        Global.user_home       = None
+        Global.launcher_home   = None
+        Global.istep           = 0
+        Global.unzipped_folder = None
+        Global.git_commit_id   = None
+        Global.dependencies_ok = False
+        Global.zip_folder      = None
+        Global.install_step_ok = {}
+        return
+    try:
+        d={}
+        d=pickle.load(open(GLOBAL_PICKLED))
+        for k,v in d.iteritems():
+            setattr(Global, k, v)
+    except:
+        loadGlobal(reset=True)
+        
+def dumpGlobal():
+    """ persistence for Global namespace"""
+    d={}
+    for k,v in Global.__dict__.iteritems():
+        if not (k.startswith('__') and k.endswith('__')):
+            d[k]=v 
+    pickle.dump(d, open(GLOBAL_PICKLED,'w+'))
+
+def removeGlobal():
+    try:
+        os.remove(GLOBAL_PICKLED)
+    except Exception as e:
+        #print type(e),e
+        pass
 
 class LogAction(object):
     def __init__(self,before='',after='done',verbose=True):
@@ -67,8 +102,6 @@ def module_version_insufficient(module_name,version,minimal_version):
                 , minimal_version= tuple2version_string(minimal_version)
                 , python_exe     = sys.executable
                 )
-    global issues
-    issues += 1
     return s
 
 def check_module(module_name, minimal_version, verbose=False, module_name_verbose=None):
@@ -104,9 +137,9 @@ def check_module(module_name, minimal_version, verbose=False, module_name_verbos
     return ok
         
 def reinstall_previous_version(verbose):
-    launcher_src_previous = os.path.join(launcher_home,LAUNCHER_PREVIOUS)
+    launcher_src_previous = os.path.join(Global.launcher_home,LAUNCHER_PREVIOUS)
     if os.path.exists(launcher_src_previous):
-        launcher_src_folder = os.path.join(launcher_home,'Launcher')
+        launcher_src_folder = os.path.join(Global.launcher_home,'Launcher')
         if os.path.exists(launcher_src_folder):
             with LogAction('  + removing recently installed Launcher version ...',verbose=verbose):
                 shutil.rmtree(launcher_src_folder)
@@ -118,33 +151,36 @@ def preconditions(args):
     verbose=not args.quiet
     # existence of home directory
     env_home = "HOMEPATH" if sys.platform=="win32" else "HOME"
-    user_home = os.environ[env_home]
-    assert (not user_home is None),"Unable to detect home folder."
+    Global.user_home = os.environ[env_home]
+    assert (not Global.user_home is None),"Unable to detect home folder."
     if verbose:
-        print '  + home folder found :', user_home
+        print '  + home folder found :', Global.user_home
     # existence of Launcher directory
-    global launcher_home
-    launcher_home = os.path.join(user_home,'Launcher')
-    if not os.path.exists(launcher_home):
+    Global.launcher_home = os.path.join(Global.user_home,'Launcher')
+    if not os.path.exists(Global.launcher_home):
         try:
-            os.makedirs(launcher_home)
+            os.makedirs(Global.launcher_home)
         except:
-            print '  - Unable to create folder "'+launcher_home+'".'
+            print '  - Unable to create folder "'+Global.launcher_home+'".'
             raise
         finally:
             if verbose:
-                print '  + Folder "'+launcher_home+'" created.'
+                print '  + Folder "'+Global.launcher_home+'" created.'
     else:
         if verbose:
-            print '  + Folder "'+launcher_home+'" exists.'
+            print '  + Folder "'+Global.launcher_home+'" exists.'
 
     # If the current directory is <$home>/Launcher/Launcher move up to parent directory
     cwd = os.getcwd()
-    if cwd.endswith('/Launcher') and not cwd==launcher_home:
+    if cwd.endswith('/Launcher') and not cwd==Global.launcher_home:
         os.chdir('..')
         if verbose:
             print '  + changing current working directory to parent directory:'
             print '     ', os.getcwd()
+    if args.force:
+        removed = removeGlobal() #forget the state of the previous run
+        if verbose and removed:
+            print '  + removed old "{}" (--force):'.format(GLOBAL_PICKLED)
     
     if args.force:
         # remove files left by previous run
@@ -161,16 +197,18 @@ def preconditions(args):
                 if file.startswith(GIT_COMMIT_ID_PREFIX):
                     remove_files.append(file)
             break
-        with LogAction('  + removing folders :',verbose=verbose):
-            for f in remove_folders:
-                if verbose:
-                    print f,'..',
-                shutil.rmtree(f)
-        with LogAction('  + removing files :',verbose=verbose):
-            for f in remove_files:
-                if verbose:
-                    print f,'..',
-                os.remove(f)
+        if remove_folders:
+            with LogAction('  + removing old folders :',verbose=verbose):
+                for f in remove_folders:
+                    if verbose:
+                        print f,'..',
+                    shutil.rmtree(f)
+        if remove_files:
+            with LogAction('  + removing old files :',verbose=verbose):
+                for f in remove_files:
+                    if verbose:
+                        print f,'..',
+                    os.remove(f)
 
     reinstall_previous_version(verbose)
     
@@ -180,7 +218,7 @@ def verifyDependencies(args):
     print "Verifying dependencies ..."
     # compliance of non standard python libraries 
     verbose = not args.quiet
-    ok=True
+    ok = True
     ok = ok and check_module('wx'      ,MINIMAL_WX_VERSION      ,verbose,module_name_verbose='wxPython')
     ok = ok and check_module('paramiko',MINIMAL_PARAMIKO_VERSION,verbose)
     if ok:
@@ -199,6 +237,8 @@ def verifyDependencies(args):
                   '    You can continue the installation by providing the --ignore-deps command '\
                   '    line argument. The startup script "launcher'+ext+'" will use the current '\
                   '    non comliant Python distribution.'
+
+    Global.dependencies_ok = ok
     return ok
     
 def download(args,branch='master'):
@@ -216,6 +256,11 @@ def download(args,branch='master'):
                 if verbose:
                     print '    attempt {} ...'.format(attempts), 
                 response = urllib2.urlopen(url,timeout=5)
+#                 print response.info()['Content-Disposition']
+#                 print type(response.info()['Content-Disposition'])
+                pattern = re.compile(r"attachment; filename=(.*)")
+                m = re.match(pattern, response.info()['Content-Disposition'])
+                Global.zip_folder = m.group(1)
                 content = response.read()
                 f = open(ZIP_FILE, 'w+' )
                 f.write( content )
@@ -228,7 +273,7 @@ def download(args,branch='master'):
                 print e
             finally:
                 if verbose:
-                    print 'succeeded :',ZIP_FILE
+                    print 'succeeded :"{}" -> "{}"'.format(Global.zip_folder,ZIP_FILE)
         if not attempts<max_attempts:
             print '  - Unable to download "{}", after {} attempts.'.format(url,attempts)
             return False
@@ -247,52 +292,62 @@ def unzip(args):
         zf = zipfile.ZipFile(ZIP_FILE,mode='r')
         zf.extractall()
     finally:
-        global unzipped_folder,git_commit_id
-        for folder,sub_folders,files in os.walk('.'):
-            for sub_folder in sub_folders:
-                if sub_folder.startswith(UNZIPPED_FOLDER_PREFIX):
-                    unzipped_folder = sub_folder
-                    git_commit_id   = sub_folder[len(UNZIPPED_FOLDER_PREFIX):]
-            break
-        if unzipped_folder is None:
-            print '  - Unzipped folder not found. Expecting "{}<git_commit_id>"'.format(UNZIPPED_FOLDER_PREFIX)
-            return False
+        if Global.zip_folder is None:
+            for folder,sub_folders,files in os.walk('.'):
+                for sub_folder in sub_folders:
+                    if sub_folder.startswith(UNZIPPED_FOLDER_PREFIX):
+                        Global.unzipped_folder = sub_folder
+                break
         else:
-            if not args.quiet:
-                print '  + Unzipped folder : "{}"'.format(unzipped_folder)
-            open('git_commit_id_'+git_commit_id,'w+') #empty file to store the git commit id 
-            return True
+            Global.unzipped_folder=os.path.splitext(Global.zip_folder)[0]
+            assert os.path.exists(Global.unzipped_folder)
+            
+    if Global.unzipped_folder is None:
+        print '  - Unzipped folder not found. Expecting "{}<Global.git_commit_id>"'.format(UNZIPPED_FOLDER_PREFIX)
+        return False
+    else:
+        if not args.quiet:
+            print '  + Unzipped folder : "{}"'.format(Global.unzipped_folder)
+        Global.git_commit_id = Global.unzipped_folder[len(UNZIPPED_FOLDER_PREFIX):]
+        open('git_commit_id_'+Global.git_commit_id,'w+') #empty file to store the git commit id 
+        return True
    
 def install(args):
     print 'Installing Launcher ...'
     verbose = not args.quiet
-    launcher_src_folder = os.path.join(launcher_home,'Launcher')
+    launcher_src_folder = os.path.join(Global.launcher_home,'Launcher')
     if os.path.exists(launcher_src_folder):
         with LogAction('  + making backup of previous version of Launcher -> "{}" ...'.format(LAUNCHER_PREVIOUS),verbose=verbose):
-            launcher_src_backup = os.path.join(launcher_home,LAUNCHER_PREVIOUS)
+            launcher_src_backup = os.path.join(Global.launcher_home,LAUNCHER_PREVIOUS)
             shutil.move(launcher_src_folder,launcher_src_backup)
-       
-    with LogAction('  + moving "{}" to "{}" ...'.format(unzipped_folder,launcher_src_folder),verbose=verbose):
-        shutil.move(unzipped_folder,launcher_src_folder)
-    
-    with LogAction('  + Copying "git_commit_id_{}" ...'.format(git_commit_id),verbose=verbose):
-        shutil.copy('git_commit_id_'+git_commit_id,launcher_src_folder)
+#     print os.getcwd()
+    try:
+        with LogAction('  + moving "{}" to "{}" ...'.format(Global.unzipped_folder,launcher_src_folder),verbose=verbose):
+            shutil.move(Global.unzipped_folder,launcher_src_folder)
         
+        with LogAction('  + Copying "git_commit_id_{}" ...'.format(Global.git_commit_id),verbose=verbose):
+            shutil.copy('git_commit_id_'+Global.git_commit_id,launcher_src_folder)
+    except:
+        traceback.print_exc(file=sys.stdout)
+        print "Returning to previous version."
+        reinstall_previous_version(verbose=verbose)
+        return False
+    
     return True
 
 def create_startup_script(args):
     print 'Creating install script'
     verbose = not args.quiet
-    launcher_src_folder = os.path.join(launcher_home,'Launcher')
+    launcher_src_folder = os.path.join(Global.launcher_home,'Launcher')
     ok = True
     with LogAction('  + creating startup script ...',verbose=verbose):
         startup = 'launcher'+ ('.bat' if sys.platform=='win32' else '.sh')
-        fpath = os.path.join(launcher_home,startup)
+        fpath = os.path.join(Global.launcher_home,startup)
         f = open(fpath,'w+')
         if not sys.platform=='win32':
             f.write("#!/bin/bash\n")
-        installer_log = pickle.load(open('installer.log'))
-        if not installer_log['verifyDependencies']:
+        
+        if not Global.dependencies_ok:
             print '  - WARNING:'\
                   '    The startup script "'+startup+'" uses a non compliant Python distribution.'\
                   '    Launcher may not work properly.'
@@ -306,22 +361,17 @@ def create_startup_script(args):
     return ok
 
     
-def install_step(step,args,force_always=False):
+def install_step(step,args):
     start = datetime.datetime.now()
-    global istep
-    istep += 1
-    print '\nSTEP', istep    
+    loadGlobal()
+    
+    Global.istep += 1
+    print '\nSTEP', Global.istep    
+    
     step_name = step.__name__
-    try:
-        installer_log = pickle.load(open('installer.log'))
-    except:
-        installer_log={}
-    if force_always:
-        step_already_ok = False
-    else:
-        step_already_ok = installer_log.get(step_name,False)
+    step_already_ok = Global.install_step_ok.get(step_name,False)
 
-    if args.force or not step_already_ok:
+    if not step_already_ok:
         try:
             success = step(args)
         except Exception as e:
@@ -333,8 +383,8 @@ def install_step(step,args,force_always=False):
             + str((datetime.datetime.now()-start).total_seconds())+'s.'
         print msg
         
-        installer_log[step_name] = success
-        pickle.dump(installer_log,open('installer.log','w+'))
+        Global.install_step_ok[step_name] = success
+        dumpGlobal()
         return success
     else:
         print '  + step not repeated since already successful and --force==False.'
@@ -361,8 +411,10 @@ if __name__=="__main__":
                        , action="store_true"
                        )
     args = parser.parse_args()
-    print args
-    success = install_step( preconditions, args, force_always=True )
+    if not args.quiet:
+        print "Commandline arguments:",args
+    
+    success     = install_step( preconditions, args)
     if success:
         success = install_step( verifyDependencies, args )
     if success:
@@ -373,5 +425,11 @@ if __name__=="__main__":
         success = install_step( install, args )
     if success:
         args.force = True
-        success = install_step( create_startup_script, args )      
-    print '\nInstallation of Launcher finished! '  
+        success = install_step( create_startup_script, args )    
+    
+    if success:
+        removeGlobal() #will interfere with next update
+        print '\nInstallation of Launcher finished successfully. Enjoy!'
+    else:
+        print '\nInstallation of Launcher failed.'
+
